% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the BHAI.AS4 Shiny App}
\usage{
launch_app()
}
\description{
This function launches the interactive BHAI Shiny application included with the BHAI.AS4 package.
The app allows exploration of HAI data (cases, deaths, DALYs) with interactive plots.
}
